initial commit
Add the wise project codebase. The project at
this point is at a very early development stage
and does not support external communication only
printing detections into the console.

This adds the `rcon` crate used to communicate
with a HLL server. It can be used as a standalone
library however is developed for the `wise` project
and only supports a few commands.

`rcon` provides:
 - A standalone async `RconConnection`
 - Built-in parsing for all supported commands

It also adds the `wise` crate representing the core of the
polling logic.

`wise` provides:
 - CLI & File configuration using `serde` and `clap`
 - The `Manager` struct to centrally track and manage pollers
 - Polling for the `PlayerInfo` command
 - Polling for the `ShowLog` command